Refs #27046 - Resolv the DNS server IP if needed
This is an alternative implementation to
79ceedcee3bb3e334cea2f7219a8820ae44dc6dd.

The root cause of the issue is that if the DNS server is a hostname, the
resolver can't find it because it has no way to get the IP address of
the resolver. By using the system resolver (as the HTTP API would), we
get equivalent behavior.