Separate IPv4 and v6 handling for PTR records
Infoblox only contains a single type for PTRs, which can contain either
an IPv4 or IPv6 address. So unless limiting searches for them in some
way, then an IPv4 PTR for the same FQDN as an IPv6 PTR will end up
looking like a collision.

With an additional regexp match on the PTR record name, the search can
be limited to only finding v4 (in-addr.arpa) or v6 (ip6.arpa) records.